// microshop-ui/src/app/api/stock/model/product.ts
/**
 * MicroShop.StockAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Category } from './category';
import { ProductStatus } from './product-status';


export interface Product { 
    id?: string;
    createdAt?: string;
    updatedAt?: string | null;
    name?: string | null;
    description?: string | null;
    price?: number;
    stockCount?: number;
    pictureUrl?: string | null;
    categoryId?: number;
    category?: Category;
    sellerId?: string;
    status?: ProductStatus;
}
export namespace Product {
}